requirements.txt
absl-py==2.0.0
anndata==0.9.2
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blosc2==2.2.9
bokeh==3.2.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
colorcet==3.0.1
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
dask==2023.9.0
datashader==0.15.2
datashape==0.5.2
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.12.4
fonttools==4.42.1
fsspec==2023.9.0
fuzzywuzzy==0.18.0
glmpca==0.1.0
google-auth==2.23.2
google-auth-oauthlib==1.0.0
grpcio==1.59.0
h5py==3.9.0
holoviews==1.17.1
huggingface-hub==0.17.3
idna==3.4
igraph==0.10.6
imageio==2.31.2
importlib-metadata==6.8.0
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
kiwisolver==1.4.5
lazy_loader==0.3
leidenalg==0.10.1
linkify-it-py==2.0.2
llvmlite==0.40.1
lmdb==1.4.1
locket==1.0.0
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.1
mpmath==1.3.0
msgpack==1.0.7
multipledispatch==1.0.0
natsort==8.4.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
ndindex==1.7
nest-asyncio==1.5.7
networkx==3.1
numba==0.57.1
numexpr==2.8.7
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.1
pandas==2.1.1
pandocfilters==1.5.0
panel==1.2.2
param==1.13.0
parso==0.8.3
partd==1.4.0
patsy==0.5.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prettytable==3.9.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyct==0.5.0
Pygments==2.16.1
pynndescent==0.5.10
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pyviz_comms==3.0.0
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.2
rpy2==3.5.14
rsa==4.9
safetensors==0.4.0
scanpy==1.9.5
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.3
seaborn==0.12.2
session-info==1.0.0
sinfo==0.3.4
six==1.16.0
soupsieve==2.5
stack-data==0.6.2
statsmodels==0.14.0
stdlib-list==0.9.0
sympy==1.12
tables==3.9.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
texttable==1.6.7
threadpoolctl==3.2.0
tifffile==2023.8.30
tinycss2==1.2.1
tokenizers==0.14.1
toolz==0.12.0
torch==2.1.0
torchtoolbox==0.1.8.2
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.34.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.1
uc-micro-py==1.0.2
umap-learn==0.5.3
urllib3==2.0.4
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==3.0.0
xarray==2023.8.0
xyzservices==2023.7.0
zipp==3.16.2
